+++ api-crm-erp/app/Http/Controllers/Proforma/ProformaController.php
@@ -36,6 +36,7 @@ class ProformaController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize("viewAny", Proforma::class);
         $search = $request->search;
         $client_segment_id = $request->client_segment_id;
         $asesor_id = $request->asesor_id;
@@ -46,10 +47,18 @@ class ProformaController extends Controller
         $end_date = $request->end_date;
         $state_proforma = $request->state_proforma;
 
-        $proformas = Proforma::filterAdvance($search,$client_segment_id,$asesor_id,
-                                    $product_categorie_id,$search_client,$search_product,
-                                    $start_date,$end_date,$state_proforma)
-                                    ->orderBy("id","desc")->paginate(25);
+        $proformas = Proforma::filterAdvance(
+            $search,
+            $client_segment_id,
+            $asesor_id,
+            $product_categorie_id,
+            $search_client,
+            $search_product,
+            $start_date,
+            $end_date,
+            $state_proforma
+        )
+            ->orderBy("id", "desc")->paginate(25);
 
         return response()->json([
             "total" => $proformas->total(),
@@ -57,35 +66,39 @@ class ProformaController extends Controller
         ]);
     }
 
-    public function export_proforma_general(Request $request){
-        return Excel::download(new ProformaGeneralExport($request),'proformas'.uniqid().'.xlsx');
+    public function export_proforma_general(Request $request)
+    {
+        return Excel::download(new ProformaGeneralExport($request), 'proformas' . uniqid() . '.xlsx');
     }
 
-    public function export_proforma_details(Request $request){
-        return Excel::download(new ProformaDetailExport($request),'proformas'.uniqid().'.xlsx');
+    public function export_proforma_details(Request $request)
+    {
+        return Excel::download(new ProformaDetailExport($request), 'proformas' . uniqid() . '.xlsx');
     }
 
-    public function proforma_pdf($id){
-        
+    public function proforma_pdf($id)
+    {
+
         $proforma = Proforma::findOrFail($id);
 
-        $pdf = PDF::loadView("proforma.proforma_pdf",compact("proforma"));
+        $pdf = PDF::loadView("proforma.proforma_pdf", compact("proforma"));
 
-        return $pdf->stream("proforma".$proforma->id.'-'.uniqid().'.pdf');
+        return $pdf->stream("proforma" . $proforma->id . '-' . uniqid() . '.pdf');
     }
 
-    public function search_clients(Request $request){
+    public function search_clients(Request $request)
+    {
         $n_document = $request->get("n_document");
         $full_name = $request->get("full_name");
         $phone = $request->get("phone");
 
-        $clients = Client::filterProforma($n_document,$full_name,$phone)->where("state",1)->orderBy("id","desc")->get();
+        $clients = Client::filterProforma($n_document, $full_name, $phone)->where("state", 1)->orderBy("id", "desc")->get();
 
         return response()->json([
-            "clients" => $clients->map(function($client){
+            "clients" => $clients->map(function ($client) {
                 return [
                     "id" => $client->id,
-                    "full_name" =>  $client->full_name,
+                    "full_name" => $client->full_name,
                     "client_segment" => $client->client_segment,
                     "phone" => $client->phone,
                     "type" => $client->type,
@@ -95,44 +108,46 @@ class ProformaController extends Controller
             })
         ]);
     }
-    public function search_products(Request $request){
+    public function search_products(Request $request)
+    {
         $search = $request->get("search");
-        $products = Product::where(DB::raw("CONCAT(products.title,' ',products.sku)"),"like","%".$search."%")->orderBy("id","desc")->get();
+        $products = Product::where(DB::raw("CONCAT(products.title,' ',products.sku)"), "like", "%" . $search . "%")->orderBy("id", "desc")->get();
 
         return response()->json([
             "products" => ProductCollection::make($products),
         ]);
     }
-    public function config(){
-        $client_segments = ClientSegment::where("state",1)->get();
-        $asesores = User::whereHas("roles",function($q) {
-            $q->where("name","like","%asesor%");
+    public function config()
+    {
+        $client_segments = ClientSegment::where("state", 1)->get();
+        $asesores = User::whereHas("roles", function ($q) {
+            $q->where("name", "like", "%asesor%");
         })->get();//->where("state",1)
-        $sucursale_deliveries = SucursaleDeliverie::where("state",1)->get();
-        $method_payments = MethodPayment::where("state",1)->where("method_payment_id",NULL)->get();
+        $sucursale_deliveries = SucursaleDeliverie::where("state", 1)->get();
+        $method_payments = MethodPayment::where("state", 1)->where("method_payment_id", NULL)->get();
         date_default_timezone_set("America/Lima");
         $today = now()->format("Y/m/d");
-        $product_categories = ProductCategorie::where("state",1)->get();
+        $product_categories = ProductCategorie::where("state", 1)->get();
         return response()->json([
             "client_segments" => $client_segments,
             "product_categories" => $product_categories,
-            "asesores" => $asesores->map(function($user) {
+            "asesores" => $asesores->map(function ($user) {
                 return [
                     "id" => $user->id,
-                    "full_name" => $user->name.' '.$user->surname,
+                    "full_name" => $user->name . ' ' . $user->surname,
                 ];
             }),
-            "sucursale_deliveries" => $sucursale_deliveries->map(function($sucursale_del) {
+            "sucursale_deliveries" => $sucursale_deliveries->map(function ($sucursale_del) {
                 return [
                     "id" => $sucursale_del->id,
                     "name" => $sucursale_del->name,
                 ];
             }),
-            "method_payments" => $method_payments->map(function($method_payment) {
+            "method_payments" => $method_payments->map(function ($method_payment) {
                 return [
                     "id" => $method_payment->id,
                     "name" => $method_payment->name,
-                    "bancos" => $method_payment->method_payments->map(function($children) {
+                    "bancos" => $method_payment->method_payments->map(function ($children) {
                         return [
                             "id" => $children->id,
                             "name" => $children->name,
@@ -144,7 +159,8 @@ class ProformaController extends Controller
         ]);
     }
 
-    public function eval_disponibilidad(Request $request,$id){
+    public function eval_disponibilidad(Request $request, $id)
+    {
         $unit_id = $request->get("unit_id");
         $previo_count_product = $request->get("quantity");
         $product_id = $id;
@@ -154,30 +170,30 @@ class ProformaController extends Controller
 
         $product = Product::findOrFail($product_id);
         // CANTIDAD DE PRODUCTOS PREVISTO PARA ATENDER O DESPACHAR
-        $count_product = ProformaDetail::where("product_id",$product_id)
-                                    ->where("unit_id",$unit_id)
-                                    ->whereHas("proforma",function($q) {
-                                        $q->where("state_proforma",2)
-                                        ->where("state_despacho",1);
-                                    })
-                                    ->sum("quantity");
-        
-        $warehouse = Warehouse::where("name","like","%".$sucursale_name."%")->first();
-        if(!$warehouse){
+        $count_product = ProformaDetail::where("product_id", $product_id)
+            ->where("unit_id", $unit_id)
+            ->whereHas("proforma", function ($q) {
+                $q->where("state_proforma", 2)
+                    ->where("state_despacho", 1);
+            })
+            ->sum("quantity");
+
+        $warehouse = Warehouse::where("name", "like", "%" . $sucursale_name . "%")->first();
+        if (!$warehouse) {
             return response()->json([
                 "message" => "No se reconocio el almacen con la sucursale de origen",
             ]);
         }
         $warehouse_id = $warehouse->id;
-        $product_warehouse = ProductWarehouse::where("product_id",$product_id)
-                          ->where("unit_id",$unit_id)
-                          ->where("warehouse_id",$warehouse_id)
-                          ->first();
-        if(!$product_warehouse){
-            $message = "El producto NO ESTA PARA disponibilidad imediata de atención y el tiempo de abastecimiento es ( ".($product->tiempo_de_abastecimiento+1)." ) dias";
+        $product_warehouse = ProductWarehouse::where("product_id", $product_id)
+            ->where("unit_id", $unit_id)
+            ->where("warehouse_id", $warehouse_id)
+            ->first();
+        if (!$product_warehouse) {
+            $message = "El producto NO ESTA PARA disponibilidad imediata de atención y el tiempo de abastecimiento es ( " . ($product->tiempo_de_abastecimiento + 1) . " ) dias";
         }
-        if($product_warehouse && $product_warehouse->stock < ($count_product + $previo_count_product)){
-            $message = "El producto no cuenta con una disponibilidad imediata de atención y el tiempo de abastecimiento es ( ".$product->tiempo_de_abastecimiento." ) dias";
+        if ($product_warehouse && $product_warehouse->stock < ($count_product + $previo_count_product)) {
+            $message = "El producto no cuenta con una disponibilidad imediata de atención y el tiempo de abastecimiento es ( " . $product->tiempo_de_abastecimiento . " ) dias";
         }
         return response()->json([
             "message" => $message,
@@ -188,6 +204,7 @@ class ProformaController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize("create", Proforma::class);
         try {
             DB::beginTransaction();
             $client = Client::findOrFail($request->client_id);
@@ -205,9 +222,9 @@ class ProformaController extends Controller
                 "description" => $request->description,
                 "state_proforma" => $client->is_parcial == 2 ? 2 : 1,
             ]);
-    
-            $DETAIL_PROFORMAS = json_decode($request->DETAIL_PROFORMAS,true);
-    
+
+            $DETAIL_PROFORMAS = json_decode($request->DETAIL_PROFORMAS, true);
+
             foreach ($DETAIL_PROFORMAS as $key => $DETAIL) {
                 ProformaDetail::create([
                     "proforma_id" => $proforma->id,
@@ -223,8 +240,8 @@ class ProformaController extends Controller
                     "total" => $DETAIL["total"],
                 ]);
             }
-    
-          ProformaDeliverie::create([
+
+            ProformaDeliverie::create([
                 "proforma_id" => $proforma->id,
                 "sucursale_deliverie_id" => $request->sucursale_deliverie_id,
                 "date_entrega" => $request->date_entrega,
@@ -240,11 +257,11 @@ class ProformaController extends Controller
                 "telefono_encargado" => $request->telefono_encargado,
             ]);
             $vaucher = "";
-    
-            if($request->hasFile("payment_file")){
-                $vaucher = Storage::putFile("proformas",$request->file("payment_file"));
+
+            if ($request->hasFile("payment_file")) {
+                $vaucher = Storage::putFile("proformas", $request->file("payment_file"));
             }
-            if($request->method_payment_id){
+            if ($request->method_payment_id) {
                 ProformaPayment::create([
                     "proforma_id" => $proforma->id,
                     "method_payment_id" => $request->method_payment_id,
@@ -256,7 +273,7 @@ class ProformaController extends Controller
             DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
-            throw new HttpException(500,$th->getMessage());
+            throw new HttpException(500, $th->getMessage());
         }
 
         return response()->json([
@@ -269,11 +286,12 @@ class ProformaController extends Controller
      */
     public function show(string $id)
     {
-       $proforma = Proforma::findOrFail($id);
+        $this->authorize("view", Proforma::class);
+        $proforma = Proforma::findOrFail($id);
 
-       return response()->json([
-        "proforma" => ProformaResource::make($proforma),
-       ]);
+        return response()->json([
+            "proforma" => ProformaResource::make($proforma),
+        ]);
     }
 
     /**
@@ -281,6 +299,7 @@ class ProformaController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $this->authorize("update", Proforma::class);
         $proforma = Proforma::findOrFail($id);
         $proforma->update($request->all());
         $proforma->proforma_deliverie->update([
@@ -307,6 +326,7 @@ class ProformaController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize("delete", Proforma::class);
         $proforma = Proforma::findOrFail($id);
         // VALIDACION POR PROFORMA
         $proforma->delete();
