+++ api-crm-erp/app/Http/Controllers/RolePermissionController.php
@@ -4,7 +4,6 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Role;
-use function Laravel\Prompts\search;
 
 class RolePermissionController extends Controller
 {
@@ -13,20 +12,19 @@ class RolePermissionController extends Controller
      */
     public function index(Request $request)
     {
-        // 
+        $this->authorize("viewAny",Role::class);
         $search = $request->get("search");
 
-        $roles = Role::with(["permissions"])->where("name", "like", "%" . $search . "%")->orderBy("id", "desc")->paginate(25);
-
+        $roles = Role::with(["permissions"])->where("name","like","%".$search."%")->orderBy("id","desc")->paginate(25);
+    
         return response()->json([
             "total" => $roles->total(),
-            "roles" => collect($roles->items())->map(function ($rol) {
+            "roles" => $roles->map(function($rol) {
                 $rol->permission_pluck = $rol->permissions->pluck("name");
                 $rol->created_format_at = $rol->created_at->format("Y-m-d h:i A");
                 return $rol;
             }),
         ]);
-
     }
 
     /**
@@ -34,10 +32,9 @@ class RolePermissionController extends Controller
      */
     public function store(Request $request)
     {
-        //
-
-        $IS_ROLE = Role::where("name", $request->name)->first();
-        if ($IS_ROLE) {
+        $this->authorize("create",Role::class);
+        $IS_ROLE = Role::where("name",$request->name)->first();
+        if($IS_ROLE){
             return response()->json([
                 "message" => 403,
                 "message_text" => "EL ROL YA EXISTE"
@@ -47,18 +44,20 @@ class RolePermissionController extends Controller
             'guard_name' => 'api',
             'name' => $request->name
         ]);
-
+        // [["id" => 1,"name" => "egreso"],["id" => 2,"name" => "ingreso"],["id" => 3,"name" => "close_caja"]]
+        // ["egreso","ingreso","close_caja"]
         foreach ($request->permisions as $key => $permision) {
-            $role->givePermissionTo($permision);
+           $role->givePermissionTo($permision);
         }
+
         return response()->json([
             "message" => 200,
             "role" => [
                 "id" => $role->id,
-                "name" => $role->name,
-                "permision" => $role->permisions,
+                "permission" => $role->permissions,
                 "permission_pluck" => $role->permissions->pluck("name"),
                 "created_format_at" => $role->created_at->format("Y-m-d h:i A"),
+                "name" => $role->name,
             ]
         ]);
     }
@@ -76,9 +75,9 @@ class RolePermissionController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        //
-        $IS_ROLE = Role::where("name", $request->name)->where("id", "<>", $id)->first();
-        if ($IS_ROLE) {
+        $this->authorize("update",Role::class);
+        $IS_ROLE = Role::where("name",$request->name)->where("id","<>",$id)->first();
+        if($IS_ROLE){
             return response()->json([
                 "message" => 403,
                 "message_text" => "EL ROL YA EXISTE"
@@ -86,18 +85,18 @@ class RolePermissionController extends Controller
         }
         $role = Role::findOrFail($id);
         $role->update($request->all());
-
-
+        // [["id" => 1,"name" => "egreso"],["id" => 2,"name" => "ingreso"],["id" => 3,"name" => "close_caja"]]
+        // ["egreso","ingreso","close_caja"]
         $role->syncPermissions($request->permisions);
 
         return response()->json([
             "message" => 200,
             "role" => [
                 "id" => $role->id,
-                "name" => $role->name,
-                "permision" => $role->permisions,
+                "permission" => $role->permissions,
                 "permission_pluck" => $role->permissions->pluck("name"),
                 "created_format_at" => $role->created_at->format("Y-m-d h:i A"),
+                "name" => $role->name,
             ]
         ]);
     }
@@ -107,17 +106,13 @@ class RolePermissionController extends Controller
      */
     public function destroy(string $id)
     {
-        //
-
+        $this->authorize("delete",Role::class);
         $role = Role::findOrFail($id);
-        // Validate
+        // la validacion por usuarios
         $role->delete();
 
-        return response()->json(
-            [
-                "message" => 200,
-
-            ]
-        );
+        return response()->json([
+            "message" => 200,
+        ]);
     }
 }
